Fix warningTimeout to catch errors caused by waitForLoad: true where iframe is loaded before script is run #1476
