+++ packages/core/index.js
@@ -33,6 +33,7 @@ import {
   vInfo,
   warn,
 } from './console'
+import warnOnNoResponse from './timeout'
 import defaults from './values/defaults'
 import page from './values/page'
 import settings from './values/settings'
@@ -816,7 +817,7 @@ const filterMsg = (msg) =>
     .filter((_, index) => index !== 19)
     .join(':')
 
-function trigger(calleeMsg, msg, id, noResponseWarning) {
+function trigger(calleeMsg, msg, id) {
   function logSent(route) {
     const displayMsg = calleeMsg in INIT_EVENTS ? filterMsg(msg) : msg
     info(id, route, HIGHLIGHT, FOREGROUND, HIGHLIGHT)
@@ -852,62 +853,11 @@ function trigger(calleeMsg, msg, id, noResponseWarning) {
     postMessageToIframe()
   }
 
-  function warnOnNoResponse() {
-    function warning() {
-      if (settings[id] === undefined) return // iframe has been closed while we where waiting
-
-      const { iframe, loaded, loadErrorShown, waitForLoad } = settings[id]
-
-      const { sandbox } = iframe
-
-      const hasSandbox = typeof sandbox === 'object' && sandbox.length > 0
-
-      if (!loaded && !loadErrorShown) {
-        settings[id].loadErrorShown = true
-        advise(
-          id,
-          `<rb>No response from iFrame</>
-            
-The iframe (<i>${id}</>) has not responded within ${settings[id].warningTimeout / 1000} seconds. Check <b>@iframe-resizer/child</> package has been loaded in the iframe.
-${
-  waitForLoad
-    ? `
-The <b>waitForLoad</> option is currently set to <i>'true'</>. If the iframe loads before the JavaScript runs, this option will prevent <i>iframe-resizer</> from initialising. To disable this, set the <b>waitForLoad</> option to <i>'false'</>.  
-`
-    : ''
-}${
-            hasSandbox &&
-            !(
-              sandbox.contains('allow-scripts') &&
-              sandbox.contains('allow-same-origin')
-            )
-              ? `
-The iframe has the <b>sandbox</> attribute, please ensure it contains both the <i>'allow-same-origin'</> and <i>'allow-scripts'</> values.
-`
-              : ''
-          } 
-${
-  hasSandbox &&
-  !(sandbox.contains('allow-scripts') && sandbox.contains('allow-same-origin'))
-    ? `The iframe has the <b>sandbox</> attribute, please ensure it contains both the <i>'allow-same-origin'</> and <i>'allow-scripts'</> values.
-`
-    : ''
-}This message can be ignored if everything is working, or you can set the <b>warningTimeout</> option to a higher value or zero to suppress this warning.
-`,
-        )
-      }
-    }
-
-    if (!!noResponseWarning && !!settings[id]?.warningTimeout) {
-      settings[id].msgTimeout = setTimeout(warning, settings[id].warningTimeout)
-    }
-  }
-
   consoleEvent(id, calleeMsg)
 
   if (settings[id]) {
     checkAndSend()
-    warnOnNoResponse()
+    // if (noResponseWarning) warnOnNoResponse()
   }
 }
 
@@ -1069,18 +1019,20 @@ Use of the <b>resize()</> method from the parent page is deprecated and will be 
   // event listener also catches the page changing in the iFrame.
   function init(msg) {
     function iFrameLoaded() {
-      trigger(ONLOAD, `${msg}:${setup}`, id, true)
+      trigger(ONLOAD, `${msg}:${setup}`, id)
       checkReset()
     }
 
     const { id } = iframe
     const { mode, waitForLoad } = settings[id]
 
+    warnOnNoResponse(id, settings)
+
     if (mode === -1) return // modal()
     if (mode === -2) return
 
     addEventListener(iframe, 'load', iFrameLoaded)
-    if (waitForLoad === false) trigger(INIT, `${msg}:${setup}`, id, true)
+    if (waitForLoad === false) trigger(INIT, `${msg}:${setup}`, id)
   }
 
   function checkOptions(options) {
